backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask==3.0.3",
    "flask-cors==4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "groq>=1.2.0",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
    "python-dotenv==1.0.1",
    "werkzeug==3.0.3",
]
